refactor(ndt_omp): prefix with autoware_
