+++ README.md
@@ -18,38 +18,38 @@ single : 282.222[msec]
 10times: 2921.92[msec]
 fitness: 0.213937
 
---- pclomp::NDT (KDTREE, 1 threads) ---
+--- autoware::ndt_omp::pclomp::NDT (KDTREE, 1 threads) ---
 single : 207.697[msec]
 10times: 2059.19[msec]
 fitness: 0.213937
 
---- pclomp::NDT (DIRECT7, 1 threads) ---
+--- autoware::ndt_omp::pclomp::NDT (DIRECT7, 1 threads) ---
 single : 139.433[msec]
 10times: 1356.79[msec]
 fitness: 0.214205
 
---- pclomp::NDT (DIRECT1, 1 threads) ---
+--- autoware::ndt_omp::pclomp::NDT (DIRECT1, 1 threads) ---
 single : 34.6418[msec]
 10times: 317.03[msec]
 fitness: 0.208511
 
---- pclomp::NDT (KDTREE, 8 threads) ---
+--- autoware::ndt_omp::pclomp::NDT (KDTREE, 8 threads) ---
 single : 54.9903[msec]
 10times: 500.51[msec]
 fitness: 0.213937
 
---- pclomp::NDT (DIRECT7, 8 threads) ---
+--- autoware::ndt_omp::pclomp::NDT (DIRECT7, 8 threads) ---
 single : 63.1442[msec]
 10times: 343.336[msec]
 fitness: 0.214205
 
---- pclomp::NDT (DIRECT1, 8 threads) ---
+--- autoware::ndt_omp::pclomp::NDT (DIRECT1, 8 threads) ---
 single : 17.2353[msec]
 10times: 100.025[msec]
 fitness: 0.208511
 ```
 
-Several methods for neighbor voxel search are implemented. If you select pclomp::KDTREE, results will be completely same as the original pcl::NDT. We recommend to use pclomp::DIRECT7 which is faster and stable. If you need extremely fast registration, choose pclomp::DIRECT1, but it might be a bit unstable.
+Several methods for neighbor voxel search are implemented. If you select autoware::ndt_omp::pclomp::KDTREE, results will be completely same as the original pcl::NDT. We recommend to use autoware::ndt_omp::pclomp::DIRECT7 which is faster and stable. If you need extremely fast registration, choose autoware::ndt_omp::pclomp::DIRECT1, but it might be a bit unstable.
 
 <img src="data/screenshot.png" height="400pix" /><br>
 Red: target, Green: source, Blue: aligned
