+++ include/autoware/ndt_omp/multigrid_pclomp/multi_voxel_grid_covariance_omp.h
@@ -69,7 +69,7 @@
 #include <map>
 #include <unordered_map>
 
-namespace pclomp
+namespace autoware::ndt_omp::pclomp
 {
 /** \brief A searchable voxel structure containing the mean and covariance of the data.
  * \note For more information please see
@@ -182,43 +182,43 @@ public:
     /** \brief Get the voxel covariance.
      * \return covariance matrix
      */
-    const Eigen::Matrix3d & getCov() const { return (cov_); }
-    Eigen::Matrix3d & getCov() { return (cov_); }
+    const Eigen::Matrix3d & getCov() const { return cov_; }
+    Eigen::Matrix3d & getCov() { return cov_; }
 
     /** \brief Get the inverse of the voxel covariance.
      * \return inverse covariance matrix
      */
-    const Eigen::Matrix3d & getInverseCov() const { return (icov_); }
+    const Eigen::Matrix3d & getInverseCov() const { return icov_; }
 
-    Eigen::Matrix3d & getInverseCov() { return (icov_); }
+    Eigen::Matrix3d & getInverseCov() { return icov_; }
 
     /** \brief Get the voxel centroid.
      * \return centroid
      */
-    const Eigen::Vector3d & getMean() const { return (mean_); }
+    const Eigen::Vector3d & getMean() const { return mean_; }
 
-    Eigen::Vector3d & getMean() { return (mean_); }
+    Eigen::Vector3d & getMean() { return mean_; }
 
     /** \brief Get the eigen vectors of the voxel covariance.
      * \note Order corresponds with \ref getEvals
      * \return matrix whose columns contain eigen vectors
      */
-    const Eigen::Matrix3d & getEvecs() const { return (evecs_); }
+    const Eigen::Matrix3d & getEvecs() const { return evecs_; }
 
-    Eigen::Matrix3d & getEvecs() { return (evecs_); }
+    Eigen::Matrix3d & getEvecs() { return evecs_; }
 
     /** \brief Get the eigen values of the voxel covariance.
      * \note Order corresponds with \ref getEvecs
      * \return vector of eigen values
      */
-    const Eigen::Vector3d & getEvals() const { return (evals_); }
+    const Eigen::Vector3d & getEvals() const { return evals_; }
 
-    Eigen::Vector3d & getEvals() { return (evals_); }
+    Eigen::Vector3d & getEvals() { return evals_; }
 
     /** \brief Get the number of points contained by this voxel.
      * \return number of points
      */
-    int getPointCount() const { return (nr_points_); }
+    int getPointCount() const { return nr_points_; }
 
     /** \brief Number of points contained by voxel */
     int nr_points_;
@@ -429,6 +429,6 @@ protected:
   // To access leaf by the search results by kdtree
   std::vector<LeafConstPtr> leaf_ptrs_;
 };
-}  // namespace pclomp
+}  // namespace autoware::ndt_omp::pclomp
 
 #endif  // #ifndef PCL_MULTI_VOXEL_GRID_COVARIANCE_H_
