+++ include/autoware/ndt_omp/multigrid_pclomp/multigrid_ndt_omp.h
@@ -56,7 +56,7 @@
 #define PCL_REGISTRATION_NDT_OMP_MULTI_VOXEL_H_
 
 #include "../pclomp/ndt_struct.hpp"
-#include "multigrid_pclomp/multi_voxel_grid_covariance_omp.h"
+#include "autoware/ndt_omp/multigrid_pclomp/multi_voxel_grid_covariance_omp.h"
 
 #include <pcl/search/impl/search.hpp>
 #include <unsupported/Eigen/NonLinearOptimization>
@@ -65,7 +65,7 @@
 
 #include <pcl/registration/registration.h>
 
-namespace pclomp
+namespace autoware::ndt_omp::pclomp
 {
 
 /** \brief A 3D Normal Distribution Transform registration implementation for point cloud data.
@@ -93,7 +93,7 @@ protected:
   typedef typename PointCloudTarget::ConstPtr PointCloudTargetConstPtr;
 
   /** \brief Typename of searchable voxel grid containing mean and covariance. */
-  typedef pclomp::MultiVoxelGridCovariance<PointTarget> TargetGrid;
+  typedef autoware::ndt_omp::pclomp::MultiVoxelGridCovariance<PointTarget> TargetGrid;
   /** \brief Typename of const pointer to searchable voxel grid leaf. */
   typedef typename TargetGrid::LeafConstPtr TargetGridLeafConstPtr;
 
@@ -181,19 +181,21 @@ public:
     // Prevents unnecessary voxel initiations
     if (params_.resolution != resolution) {
       params_.resolution = resolution;
-      if (input_) init();
+      if (input_) {
+        init();
+      }
     }
   }
 
   /** \brief Get voxel grid resolution.
    * \return side length of voxels
    */
-  inline float getResolution() const { return (params_.resolution); }
+  inline float getResolution() const { return params_.resolution; }
 
   /** \brief Get the newton line search maximum step length.
    * \return maximum step length
    */
-  inline double getStepSize() const { return (params_.step_size); }
+  inline double getStepSize() const { return params_.step_size; }
 
   /** \brief Set/change the newton line search maximum step length.
    * \param[in] step_size maximum step length
@@ -203,7 +205,7 @@ public:
   /** \brief Get the point cloud outlier ratio.
    * \return outlier ratio
    */
-  inline double getOutlierRatio() const { return (outlier_ratio_); }
+  inline double getOutlierRatio() const { return outlier_ratio_; }
 
   /** \brief Set/change the point cloud outlier ratio.
    * \param[in] outlier_ratio outlier ratio
@@ -213,7 +215,7 @@ public:
   /** \brief Get the registration alignment probability.
    * \return transformation probability
    */
-  inline double getTransformationProbability() const { return (trans_probability_); }
+  inline double getTransformationProbability() const { return trans_probability_; }
 
   inline double getNearestVoxelTransformationLikelihood() const
   {
@@ -223,7 +225,7 @@ public:
   /** \brief Get the number of iterations required to calculate alignment.
    * \return final number of iterations
    */
-  inline int getFinalNumIteration() const { return (nr_iterations_); }
+  inline int getFinalNumIteration() const { return nr_iterations_; }
 
   /** \brief Return the hessian matrix */
   inline Eigen::Matrix<double, 6, 6> getHessian() const { return hessian_; }
@@ -303,7 +305,7 @@ public:
   /** \brief Get the transformation epsilon (maximum allowable translation squared
    * difference between two consecutive transformations) as set by the user.
    */
-  inline double getTransformationEpsilon() { return (params_.trans_epsilon); }
+  inline double getTransformationEpsilon() { return params_.trans_epsilon; }
 
   inline void setMaximumIterations(int max_iterations)
   {
@@ -501,7 +503,7 @@ protected:
   inline double auxiliaryFunction_PsiMT(
     double a, double f_a, double f_0, double g_0, double mu = 1.e-4)
   {
-    return (f_a - f_0 - mu * g_0 * a);
+    return f_a - f_0 - mu * g_0 * a;
   }
 
   /** \brief Auxiliary function derivative used to determine endpoints of More-Thuente interval.
@@ -513,7 +515,7 @@ protected:
    */
   inline double auxiliaryFunction_dPsiMT(double g_a, double g_0, double mu = 1.e-4)
   {
-    return (g_a - mu * g_0);
+    return g_a - mu * g_0;
   }
 
   /** \brief The voxel grid generated from target cloud containing point means and covariances. */
@@ -562,6 +564,6 @@ public:
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
 };
 
-}  // namespace pclomp
+}  // namespace autoware::ndt_omp::pclomp
 
 #endif  // PCL_REGISTRATION_NDT_MULTI_VOXEL_H_
