+++ include/autoware/ndt_omp/multigrid_pclomp/multigrid_ndt_omp_impl.hpp
@@ -1,3 +1,17 @@
+// Copyright 2022 TIER IV, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 /*
  * Software License Agreement (BSD License)
  *
@@ -38,18 +52,129 @@
  *
  */
 
-#ifndef PCL_REGISTRATION_NDT_OMP_IMPL_H_
-#define PCL_REGISTRATION_NDT_OMP_IMPL_H_
+#ifndef PCL_REGISTRATION_NDT_OMP_MULTI_VOXEL_IMPL_H_
+#define PCL_REGISTRATION_NDT_OMP_MULTI_VOXEL_IMPL_H_
 
-#include "ndt_omp.h"
+#include "autoware/ndt_omp/multigrid_pclomp/multigrid_ndt_omp.h"
 
 #include <algorithm>
-#include <cmath>
+#include <utility>
 #include <vector>
+
+namespace autoware::ndt_omp::pclomp
+{
+
+template <typename PointSource, typename PointTarget>
+MultiGridNormalDistributionsTransform<PointSource, PointTarget>::
+  MultiGridNormalDistributionsTransform(const MultiGridNormalDistributionsTransform & other)
+: BaseRegType(other), target_cells_(other.target_cells_)
+{
+  params_ = other.params_;
+  outlier_ratio_ = other.outlier_ratio_;
+  gauss_d1_ = other.gauss_d1_;
+  gauss_d2_ = other.gauss_d2_;
+  gauss_d3_ = other.gauss_d3_;
+  trans_probability_ = other.trans_probability_;
+  // No need to copy j_ang_ and h_ang_, as those matrices are re-computed on every
+  // computeDerivatives() call
+
+  hessian_ = other.hessian_;
+  transformation_array_ = other.transformation_array_;
+  transform_probability_array_ = other.transform_probability_array_;
+  nearest_voxel_transformation_likelihood_array_ =
+    other.nearest_voxel_transformation_likelihood_array_;
+  nearest_voxel_transformation_likelihood_ = other.nearest_voxel_transformation_likelihood_;
+
+  regularization_pose_ = other.regularization_pose_;
+  regularization_pose_translation_ = other.regularization_pose_translation_;
+}
+
+template <typename PointSource, typename PointTarget>
+MultiGridNormalDistributionsTransform<PointSource, PointTarget>::
+  MultiGridNormalDistributionsTransform(MultiGridNormalDistributionsTransform && other) noexcept
+: BaseRegType(std::move(other)),
+  target_cells_(std::move(other.target_cells_)),
+  params_(std::move(other.params_))
+{
+  outlier_ratio_ = other.outlier_ratio_;
+  gauss_d1_ = other.gauss_d1_;
+  gauss_d2_ = other.gauss_d2_;
+  gauss_d3_ = other.gauss_d3_;
+  trans_probability_ = other.trans_probability_;
+
+  hessian_ = other.hessian_;
+  transformation_array_ = other.transformation_array_;
+  transform_probability_array_ = other.transform_probability_array_;
+  nearest_voxel_transformation_likelihood_array_ =
+    other.nearest_voxel_transformation_likelihood_array_;
+  nearest_voxel_transformation_likelihood_ = other.nearest_voxel_transformation_likelihood_;
+
+  regularization_pose_ = other.regularization_pose_;
+  regularization_pose_translation_ = other.regularization_pose_translation_;
+}
+
+template <typename PointSource, typename PointTarget>
+MultiGridNormalDistributionsTransform<PointSource, PointTarget> &
+MultiGridNormalDistributionsTransform<PointSource, PointTarget>::operator=(
+  const MultiGridNormalDistributionsTransform & other)
+{
+  target_cells_ = other.target_cells_;
+  params_ = other.params_;
+
+  outlier_ratio_ = other.outlier_ratio_;
+  gauss_d1_ = other.gauss_d1_;
+  gauss_d2_ = other.gauss_d2_;
+  gauss_d3_ = other.gauss_d3_;
+  trans_probability_ = other.trans_probability_;
+
+  hessian_ = other.hessian_;
+  transformation_array_ = other.transformation_array_;
+  transform_probability_array_ = other.transform_probability_array_;
+  nearest_voxel_transformation_likelihood_array_ =
+    other.nearest_voxel_transformation_likelihood_array_;
+  nearest_voxel_transformation_likelihood_ = other.nearest_voxel_transformation_likelihood_;
+
+  regularization_pose_ = other.regularization_pose_;
+  regularization_pose_translation_ = other.regularization_pose_translation_;
+
+  BaseRegType::operator=(other);
+
+  return *this;
+}
+
+template <typename PointSource, typename PointTarget>
+MultiGridNormalDistributionsTransform<PointSource, PointTarget> &
+MultiGridNormalDistributionsTransform<PointSource, PointTarget>::operator=(
+  MultiGridNormalDistributionsTransform && other) noexcept
+{
+  target_cells_ = std::move(other.target_cells_);
+  params_ = std::move(other.params_);
+
+  outlier_ratio_ = other.outlier_ratio_;
+  gauss_d1_ = other.gauss_d1_;
+  gauss_d2_ = other.gauss_d2_;
+  gauss_d3_ = other.gauss_d3_;
+  trans_probability_ = other.trans_probability_;
+
+  hessian_ = other.hessian_;
+  transformation_array_ = other.transformation_array_;
+  transform_probability_array_ = other.transform_probability_array_;
+  nearest_voxel_transformation_likelihood_array_ =
+    other.nearest_voxel_transformation_likelihood_array_;
+  nearest_voxel_transformation_likelihood_ = other.nearest_voxel_transformation_likelihood_;
+
+  regularization_pose_ = other.regularization_pose_;
+  regularization_pose_translation_ = other.regularization_pose_translation_;
+
+  BaseRegType::operator=(std::move(other));
+
+  return *this;
+}
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-pclomp::NormalDistributionsTransform<PointSource, PointTarget>::NormalDistributionsTransform()
+MultiGridNormalDistributionsTransform<
+  PointSource, PointTarget>::MultiGridNormalDistributionsTransform()
 : target_cells_(),
   outlier_ratio_(0.55),
   gauss_d1_(),
@@ -58,23 +183,20 @@ pclomp::NormalDistributionsTransform<PointSource, PointTarget>::NormalDistributi
   trans_probability_(),
   regularization_pose_(boost::none)
 {
-  reg_name_ = "NormalDistributionsTransform";
+  reg_name_ = "MultiGridNormalDistributionsTransform";
 
   params_.trans_epsilon = 0.1;
   params_.step_size = 0.1;
   params_.resolution = 1.0f;
   params_.max_iterations = 35;
-  params_.search_method = DIRECT7;
+  params_.search_method = KDTREE;  // Only KDTREE is supported in multigrid_ndt_omp
   params_.num_threads = omp_get_max_threads();
   params_.regularization_scale_factor = 0.0f;
   params_.use_line_search = false;
 
-  double gauss_c1 = NAN;
-  double gauss_c2 = NAN;
-
   // Initializes the gaussian fitting parameters (eq. 6.8) [Magnusson 2009]
-  gauss_c1 = 10.0 * (1 - outlier_ratio_);
-  gauss_c2 = outlier_ratio_ / pow(params_.resolution, 3);
+  double gauss_c1 = 10.0 * (1 - outlier_ratio_);
+  double gauss_c2 = outlier_ratio_ / pow(params_.resolution, 3);
   gauss_d3_ = -log(gauss_c2);
   gauss_d1_ = -log(gauss_c1 + gauss_c2) - gauss_d3_;
   gauss_d2_ = -2 * log((-log(gauss_c1 * exp(-0.5) + gauss_c2) - gauss_d3_) / gauss_d1_);
@@ -82,33 +204,34 @@ pclomp::NormalDistributionsTransform<PointSource, PointTarget>::NormalDistributi
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeTransformation(
+void MultiGridNormalDistributionsTransform<PointSource, PointTarget>::computeTransformation(
   PointCloudSource & output, const Eigen::Matrix4f & guess)
 {
   nr_iterations_ = 0;
   converged_ = false;
 
-  double gauss_c1 = NAN;
-  double gauss_c2 = NAN;
-
   // Initializes the gaussian fitting parameters (eq. 6.8) [Magnusson 2009]
-  gauss_c1 = 10 * (1 - outlier_ratio_);
-  gauss_c2 = outlier_ratio_ / pow(params_.resolution, 3);
+  double gauss_c1 = 10 * (1 - outlier_ratio_);
+  double gauss_c2 = outlier_ratio_ / pow(params_.resolution, 3);
   gauss_d3_ = -log(gauss_c2);
   gauss_d1_ = -log(gauss_c1 + gauss_c2) - gauss_d3_;
-  gauss_d2_ = -2 * log((-log(gauss_c1 * exp(-0.5) + gauss_c2) - gauss_d3_) / gauss_d1_);
+  gauss_d2_ = -2.0 * log((-log(gauss_c1 * exp(-0.5) + gauss_c2) - gauss_d3_) / gauss_d1_);
 
   if (guess != Eigen::Matrix4f::Identity()) {
     // Initialise final transformation to the guessed one
     final_transformation_ = guess;
     // Apply guessed transformation prior to search for neighbours
-    transformPointCloud(*input_, output, guess);
+    transformPointCloud(output, output, guess);
   }
 
   Eigen::Transform<float, 3, Eigen::Affine, Eigen::ColMajor> eig_transformation;
+
   eig_transformation.matrix() = final_transformation_;
   transformation_array_.clear();
   transformation_array_.push_back(final_transformation_);
+
+  transform_probability_array_.clear();
+  nearest_voxel_transformation_likelihood_array_.clear();
 
   // Convert initial guess matrix to 6 element transformation vector
   Eigen::Matrix<double, 6, 1> p;
@@ -116,13 +239,13 @@ void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeTran
   Eigen::Matrix<double, 6, 1> score_gradient;
   Eigen::Vector3f init_translation = eig_transformation.translation();
   Eigen::Vector3f init_rotation = eig_transformation.rotation().eulerAngles(0, 1, 2);
+
   p << init_translation(0), init_translation(1), init_translation(2), init_rotation(0),
     init_rotation(1), init_rotation(2);
 
   Eigen::Matrix<double, 6, 6> hessian;
 
   double score = 0;
-  double delta_p_norm = NAN;
 
   if (regularization_pose_) {
     Eigen::Transform<float, 3, Eigen::Affine, Eigen::ColMajor> regularization_pose_transformation;
@@ -145,13 +268,13 @@ void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeTran
     delta_p = sv.solve(-score_gradient);
 
     // Calculate step length with guaranteed sufficient decrease [More, Thuente 1994]
-    delta_p_norm = delta_p.norm();
+    double delta_p_norm = delta_p.norm();
 
     if (delta_p_norm == 0 || delta_p_norm != delta_p_norm) {
-      if (input_->points.empty()) {
+      if (input_->empty()) {
         trans_probability_ = 0.0f;
       } else {
-        trans_probability_ = score / static_cast<double>(input_->points.size());
+        trans_probability_ = score / static_cast<double>(input_->size());
       }
 
       converged_ = delta_p_norm == delta_p_norm;
@@ -160,8 +283,8 @@ void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeTran
 
     delta_p.normalize();
     delta_p_norm = computeStepLengthMT(
-      p, delta_p, delta_p_norm, params_.step_size, params_.trans_epsilon / 2, score, score_gradient,
-      hessian, output);
+      p, delta_p, delta_p_norm, params_.step_size, params_.trans_epsilon / 2.0, score,
+      score_gradient, hessian, output);
     delta_p *= delta_p_norm;
 
     transformation_ =
@@ -178,24 +301,25 @@ void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeTran
     p = p + delta_p;
 
     // Update Visualizer (untested)
-    if (update_visualizer_ != 0)
+    if (update_visualizer_ != 0) {
       update_visualizer_(output, std::vector<int>(), *target_, std::vector<int>());
-
-    if (
-      nr_iterations_ > params_.max_iterations ||
-      (nr_iterations_ && (std::fabs(delta_p_norm) < params_.trans_epsilon))) {
-      converged_ = true;
     }
 
     nr_iterations_++;
+
+    if (
+      nr_iterations_ >= params_.max_iterations ||
+      (nr_iterations_ && (std::fabs(delta_p_norm) < params_.trans_epsilon))) {
+      converged_ = true;
+    }
   }
 
   // Store transformation probability. The relative differences within each scan registration are
   // accurate but the normalization constants need to be modified for it to be globally accurate
-  if (input_->points.empty()) {
+  if (input_->empty()) {
     trans_probability_ = 0.0f;
   } else {
-    trans_probability_ = score / static_cast<double>(input_->points.size());
+    trans_probability_ = score / static_cast<double>(input_->size());
   }
 
   hessian_ = hessian;
@@ -214,130 +338,105 @@ int omp_get_thread_num()
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeDerivatives(
+double MultiGridNormalDistributionsTransform<PointSource, PointTarget>::computeDerivatives(
   Eigen::Matrix<double, 6, 1> & score_gradient, Eigen::Matrix<double, 6, 6> & hessian,
   PointCloudSource & trans_cloud, Eigen::Matrix<double, 6, 1> & p, bool compute_hessian)
 {
   score_gradient.setZero();
   hessian.setZero();
+
   double score = 0;
   int total_neighborhood_count = 0;
   double nearest_voxel_score = 0;
   size_t found_neigborhood_voxel_num = 0;
 
-  std::vector<double> scores(input_->points.size());
-  std::vector<double> nearest_voxel_scores(input_->points.size());
-  std::vector<size_t> found_neigborhood_voxel_nums(input_->points.size());
+  std::vector<double> scores(params_.num_threads);
+  std::vector<double> nearest_voxel_scores(params_.num_threads);
+  std::vector<size_t> found_neigborhood_voxel_nums(params_.num_threads);
   std::vector<Eigen::Matrix<double, 6, 1>, Eigen::aligned_allocator<Eigen::Matrix<double, 6, 1>>>
-    score_gradients(input_->points.size());
+    score_gradients(params_.num_threads);
   std::vector<Eigen::Matrix<double, 6, 6>, Eigen::aligned_allocator<Eigen::Matrix<double, 6, 6>>>
-    hessians(input_->points.size());
-  std::vector<int> neighborhood_counts(input_->points.size());
-  for (std::size_t i = 0; i < input_->points.size(); i++) {
+    hessians(params_.num_threads);
+  std::vector<int> neighborhood_counts(params_.num_threads);
+
+  // Pre-allocate thread-wise point derivative matrices to avoid reallocate too many times
+  std::vector<Eigen::Matrix<double, 4, 6>> t_point_gradients(params_.num_threads);
+  std::vector<Eigen::Matrix<double, 24, 6>> t_point_hessians(params_.num_threads);
+
+  for (size_t i = 0; i < params_.num_threads; ++i) {
     scores[i] = 0;
     nearest_voxel_scores[i] = 0;
     found_neigborhood_voxel_nums[i] = 0;
     score_gradients[i].setZero();
     hessians[i].setZero();
     neighborhood_counts[i] = 0;
+
+    // Initialize point derivatives
+    t_point_gradients[i].setZero();
+    t_point_gradients[i].block<3, 3>(0, 0).setIdentity();
+    t_point_hessians[i].setZero();
   }
 
   // Precompute Angular Derivatives (eq. 6.19 and 6.21)[Magnusson 2009]
   computeAngleDerivatives(p);
 
-  std::vector<std::vector<TargetGridLeafConstPtr>> neighborhoods(params_.num_threads);
-  std::vector<std::vector<float>> distancess(params_.num_threads);
-
   // Update gradient and hessian for each point, line 17 in Algorithm 2 [Magnusson 2009]
 #pragma omp parallel for num_threads(params_.num_threads) schedule(guided, 8)
-  for (std::size_t idx = 0; idx < input_->points.size(); idx++) {
-    int thread_n = omp_get_thread_num();
+  for (size_t idx = 0; idx < input_->size(); ++idx) {
+    int tid = omp_get_thread_num();
+    // Searching for neighbors of the current transformed point
+    auto & x_trans_pt = trans_cloud[idx];
+    std::vector<TargetGridLeafConstPtr> neighborhood;
 
-    // Original Point and Transformed Point
-    PointSource x_pt;
-    PointSource x_trans_pt;
-    // Original Point and Transformed Point (for math)
-    Eigen::Vector3d x;
-    Eigen::Vector3d x_trans;
-    // Occupied Voxel
-    TargetGridLeafConstPtr cell;
-    // Inverse Covariance of Occupied Voxel
-    Eigen::Matrix3d c_inv;
+    // Neighborhood search method other than kdtree is disabled in multigrid_ndt_omp
+    target_cells_.radiusSearch(x_trans_pt, params_.resolution, neighborhood);
 
-    // Initialize Point Gradient and Hessian
-    Eigen::Matrix<float, 4, 6> point_gradient;
-    Eigen::Matrix<float, 24, 6> point_hessian;
-    point_gradient.setZero();
-    point_gradient.block<3, 3>(0, 0).setIdentity();
-    point_hessian.setZero();
-
-    x_trans_pt = trans_cloud.points[idx];
-
-    auto & neighborhood = neighborhoods[thread_n];
-    auto & distances = distancess[thread_n];
-
-    // Find neighbors (Radius search has been experimentally faster than direct neighbor checking.
-    switch (params_.search_method) {
-      case KDTREE:
-        target_cells_.radiusSearch(x_trans_pt, params_.resolution, neighborhood, distances);
-        break;
-      case DIRECT26:
-        target_cells_.getNeighborhoodAtPoint(x_trans_pt, neighborhood);
-        break;
-      default:
-      case DIRECT7:
-        target_cells_.getNeighborhoodAtPoint7(x_trans_pt, neighborhood);
-        break;
-      case DIRECT1:
-        target_cells_.getNeighborhoodAtPoint1(x_trans_pt, neighborhood);
-        break;
+    if (neighborhood.empty()) {
+      continue;
     }
+
+    // Original Point
+    auto & x_pt = (*input_)[idx];
+    // Original Point and Transformed Point (for math)
+    Eigen::Vector3d x(x_pt.x, x_pt.y, x_pt.z);
+    // Current Point Gradient and Hessian
+    auto & point_gradient = t_point_gradients[tid];
+    auto & point_hessian = t_point_hessians[tid];
+
+    // Compute derivative of transform function w.r.t. transform vector, J_E and H_E in
+    // Equations 6.18 and 6.20 [Magnusson 2009]
+    computePointDerivatives(x, point_gradient, point_hessian);
+
+    // Denorm point, x_k' in Equations 6.12 and 6.13 [Magnusson 2009]
+    const Eigen::Vector3d x_trans(x_trans_pt.x, x_trans_pt.y, x_trans_pt.z);
 
     double sum_score_pt = 0;
     double nearest_voxel_score_pt = 0;
-    Eigen::Matrix<double, 6, 1> score_gradient_pt = Eigen::Matrix<double, 6, 1>::Zero();
-    Eigen::Matrix<double, 6, 6> hessian_pt = Eigen::Matrix<double, 6, 6>::Zero();
-    int neighborhood_count = 0;
+    auto & score_gradient_pt = score_gradients[tid];
+    auto & hessian_pt = hessians[tid];
 
     for (auto & cell : neighborhood) {
-      x_pt = input_->points[idx];
-      x = Eigen::Vector3d(x_pt.x, x_pt.y, x_pt.z);
-
-      x_trans = Eigen::Vector3d(x_trans_pt.x, x_trans_pt.y, x_trans_pt.z);
-
-      // Denorm point, x_k' in Equations 6.12 and 6.13 [Magnusson 2009]
-      x_trans -= cell->getMean();
-      // Uses precomputed covariance for speed.
-      c_inv = cell->getInverseCov();
-
-      // Compute derivative of transform function w.r.t. transform vector, J_E and H_E in
-      // Equations 6.18 and 6.20 [Magnusson 2009]
-      computePointDerivatives(x, point_gradient, point_hessian);
       // Update score, gradient and hessian, lines 19-21 in Algorithm 2, according to
       // Equations 6.10, 6.12 and 6.13, respectively [Magnusson 2009]
       double score_pt = updateDerivatives(
-        score_gradient_pt, hessian_pt, point_gradient, point_hessian, x_trans, c_inv,
-        compute_hessian);
-      neighborhood_count++;
+        score_gradient_pt, hessian_pt, point_gradient, point_hessian, x_trans - cell->getMean(),
+        cell->getInverseCov(), compute_hessian);
       sum_score_pt += score_pt;
+
       if (score_pt > nearest_voxel_score_pt) {
         nearest_voxel_score_pt = score_pt;
       }
     }
 
-    if (!neighborhood.empty()) {
-      ++found_neigborhood_voxel_nums[idx];
-    }
+    ++found_neigborhood_voxel_nums[tid];
 
-    scores[idx] = sum_score_pt;
-    nearest_voxel_scores[idx] = nearest_voxel_score_pt;
-    score_gradients[idx].noalias() = score_gradient_pt;
-    hessians[idx].noalias() = hessian_pt;
-    neighborhood_counts[idx] += neighborhood_count;
+    scores[tid] += sum_score_pt;
+    nearest_voxel_scores[tid] += nearest_voxel_score_pt;
+    neighborhood_counts[tid] += neighborhood.size();
   }
 
   // Ensure that the result is invariant against the summing up order
-  for (std::size_t i = 0; i < input_->points.size(); i++) {
+  for (size_t i = 0; i < params_.num_threads; ++i) {
     score += scores[i];
     nearest_voxel_score += nearest_voxel_scores[i];
     found_neigborhood_voxel_num += found_neigborhood_voxel_nums[i];
@@ -387,22 +486,26 @@ double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeDe
   } else {
     nearest_voxel_transformation_likelihood_ = 0.0;
   }
-
-  return (score);
+  nearest_voxel_transformation_likelihood_array_.push_back(
+    nearest_voxel_transformation_likelihood_);
+  const float transform_probability =
+    (input_->points.empty() ? 0.0f : score / static_cast<double>(input_->points.size()));
+  transform_probability_array_.push_back(transform_probability);
+  return score;
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeAngleDerivatives(
+void MultiGridNormalDistributionsTransform<PointSource, PointTarget>::computeAngleDerivatives(
   Eigen::Matrix<double, 6, 1> & p, bool compute_hessian)
 {
   // Simplified math for near 0 angles
-  double cx = NAN;
-  double cy = NAN;
-  double cz = NAN;
-  double sx = NAN;
-  double sy = NAN;
-  double sz = NAN;
+  double cx = 1.0;
+  double cy = 1.0;
+  double cz = 1.0;
+  double sx = 0.0;
+  double sy = 0.0;
+  double sz = 0.0;
   if (fabs(p(3)) < 10e-5) {
     // p(3) = 0;
     cx = 1.0;
@@ -430,262 +533,143 @@ void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeAngl
   }
 
   // Precomputed angular gradiant components. Letters correspond to Equation 6.19 [Magnusson 2009]
-  j_ang_a_ << (-sx * sz + cx * sy * cz), (-sx * cz - cx * sy * sz), (-cx * cy);
-  j_ang_b_ << (cx * sz + sx * sy * cz), (cx * cz - sx * sy * sz), (-sx * cy);
-  j_ang_c_ << (-sy * cz), sy * sz, cy;
-  j_ang_d_ << sx * cy * cz, (-sx * cy * sz), sx * sy;
-  j_ang_e_ << (-cx * cy * cz), cx * cy * sz, (-cx * sy);
-  j_ang_f_ << (-cy * sz), (-cy * cz), 0;
-  j_ang_g_ << (cx * cz - sx * sy * sz), (-cx * sz - sx * sy * cz), 0;
-  j_ang_h_ << (sx * cz + cx * sy * sz), (cx * sy * cz - sx * sz), 0;
+  j_ang_.setZero();
 
-  j_ang.setZero();
-  j_ang.row(0).noalias() = Eigen::Vector4f(
-    static_cast<float>(-sx * sz + cx * sy * cz), static_cast<float>(-sx * cz - cx * sy * sz),
-    static_cast<float>(-cx * cy), 0.0f);
-  j_ang.row(1).noalias() = Eigen::Vector4f(
-    static_cast<float>(cx * sz + sx * sy * cz), static_cast<float>(cx * cz - sx * sy * sz),
-    static_cast<float>(-sx * cy), 0.0f);
-  j_ang.row(2).noalias() = Eigen::Vector4f(
-    static_cast<float>(-sy * cz), static_cast<float>(sy * sz), static_cast<float>(cy), 0.0f);
-  j_ang.row(3).noalias() = Eigen::Vector4f(
-    static_cast<float>(sx * cy * cz), static_cast<float>(-sx * cy * sz),
-    static_cast<float>(sx * sy), 0.0f);
-  j_ang.row(4).noalias() = Eigen::Vector4f(
-    static_cast<float>(-cx * cy * cz), static_cast<float>(cx * cy * sz),
-    static_cast<float>(-cx * sy), 0.0f);
-  j_ang.row(5).noalias() =
-    Eigen::Vector4f(static_cast<float>(-cy * sz), static_cast<float>(-cy * cz), 0.0f, 0.0f);
-  j_ang.row(6).noalias() = Eigen::Vector4f(
-    static_cast<float>(cx * cz - sx * sy * sz), static_cast<float>(-cx * sz - sx * sy * cz), 0,
-    0.0f);
-  j_ang.row(7).noalias() = Eigen::Vector4f(
-    static_cast<float>(sx * cz + cx * sy * sz), static_cast<float>(cx * sy * cz - sx * sz), 0,
-    0.0f);
+  j_ang_.row(0) << (-sx * sz + cx * sy * cz), (-sx * cz - cx * sy * sz), (-cx * cy), 0.0f;
+  j_ang_.row(1) << (cx * sz + sx * sy * cz), (cx * cz - sx * sy * sz), (-sx * cy), 0.0f;
+  j_ang_.row(2) << (-sy * cz), sy * sz, cy, 0.0f;
+  j_ang_.row(3) << sx * cy * cz, (-sx * cy * sz), sx * sy, 0.0f;
+  j_ang_.row(4) << (-cx * cy * cz), cx * cy * sz, (-cx * sy), 0.0f;
+  j_ang_.row(5) << (-cy * sz), (-cy * cz), 0.0f, 0.0f;
+  j_ang_.row(6) << (cx * cz - sx * sy * sz), (-cx * sz - sx * sy * cz), 0.0f, 0.0f;
+  j_ang_.row(7) << (sx * cz + cx * sy * sz), (cx * sy * cz - sx * sz), 0.0f, 0.0f;
 
   if (compute_hessian) {
     // Precomputed angular hessian components. Letters correspond to Equation 6.21 and numbers
     // correspond to row index [Magnusson 2009]
-    h_ang_a2_ << (-cx * sz - sx * sy * cz), (-cx * cz + sx * sy * sz), sx * cy;
-    h_ang_a3_ << (-sx * sz + cx * sy * cz), (-cx * sy * sz - sx * cz), (-cx * cy);
+    h_ang_.setZero();
 
-    h_ang_b2_ << (cx * cy * cz), (-cx * cy * sz), (cx * sy);
-    h_ang_b3_ << (sx * cy * cz), (-sx * cy * sz), (sx * sy);
+    h_ang_.row(0) << (-cx * sz - sx * sy * cz), (-cx * cz + sx * sy * sz), sx * cy, 0.0f;     // a2
+    h_ang_.row(1) << (-sx * sz + cx * sy * cz), (-cx * sy * sz - sx * cz), (-cx * cy), 0.0f;  // a3
 
-    h_ang_c2_ << (-sx * cz - cx * sy * sz), (sx * sz - cx * sy * cz), 0;
-    h_ang_c3_ << (cx * cz - sx * sy * sz), (-sx * sy * cz - cx * sz), 0;
+    h_ang_.row(2) << (cx * cy * cz), (-cx * cy * sz), (cx * sy), 0.0f;  // b2
+    h_ang_.row(3) << (sx * cy * cz), (-sx * cy * sz), (sx * sy), 0.0f;  // b3
 
-    h_ang_d1_ << (-cy * cz), (cy * sz), (sy);
-    h_ang_d2_ << (-sx * sy * cz), (sx * sy * sz), (sx * cy);
-    h_ang_d3_ << (cx * sy * cz), (-cx * sy * sz), (-cx * cy);
+    h_ang_.row(4) << (-sx * cz - cx * sy * sz), (sx * sz - cx * sy * cz), 0.0f, 0.0f;  // c2
+    h_ang_.row(5) << (cx * cz - sx * sy * sz), (-sx * sy * cz - cx * sz), 0.0f, 0.0f;  // c3
 
-    h_ang_e1_ << (sy * sz), (sy * cz), 0;
-    h_ang_e2_ << (-sx * cy * sz), (-sx * cy * cz), 0;
-    h_ang_e3_ << (cx * cy * sz), (cx * cy * cz), 0;
+    h_ang_.row(6) << (-cy * cz), (cy * sz), (sy), 0.0f;                  // d1
+    h_ang_.row(7) << (-sx * sy * cz), (sx * sy * sz), (sx * cy), 0.0f;   // d2
+    h_ang_.row(8) << (cx * sy * cz), (-cx * sy * sz), (-cx * cy), 0.0f;  // d3
 
-    h_ang_f1_ << (-cy * cz), (cy * sz), 0;
-    h_ang_f2_ << (-cx * sz - sx * sy * cz), (-cx * cz + sx * sy * sz), 0;
-    h_ang_f3_ << (-sx * sz + cx * sy * cz), (-cx * sy * sz - sx * cz), 0;
+    h_ang_.row(9) << (sy * sz), (sy * cz), 0.0f, 0.0f;               // e1
+    h_ang_.row(10) << (-sx * cy * sz), (-sx * cy * cz), 0.0f, 0.0f;  // e2
+    h_ang_.row(11) << (cx * cy * sz), (cx * cy * cz), 0.0f, 0.0f;    // e3
 
-    h_ang.setZero();
-    h_ang.row(0).noalias() = Eigen::Vector4f(
-      static_cast<float>(-cx * sz - sx * sy * cz), static_cast<float>(-cx * cz + sx * sy * sz),
-      static_cast<float>(sx * cy), 0.0f);  // a2
-    h_ang.row(1).noalias() = Eigen::Vector4f(
-      static_cast<float>(-sx * sz + cx * sy * cz), static_cast<float>(-cx * sy * sz - sx * cz),
-      static_cast<float>(-cx * cy), 0.0f);  // a3
-
-    h_ang.row(2).noalias() = Eigen::Vector4f(
-      static_cast<float>(cx * cy * cz), static_cast<float>(-cx * cy * sz),
-      static_cast<float>(cx * sy), 0.0f);  // b2
-    h_ang.row(3).noalias() = Eigen::Vector4f(
-      static_cast<float>(sx * cy * cz), static_cast<float>(-sx * cy * sz),
-      static_cast<float>(sx * sy), 0.0f);  // b3
-
-    h_ang.row(4).noalias() = Eigen::Vector4f(
-      static_cast<float>(-sx * cz - cx * sy * sz), static_cast<float>(sx * sz - cx * sy * cz), 0.0f,
-      0.0f);  // c2
-    h_ang.row(5).noalias() = Eigen::Vector4f(
-      static_cast<float>(cx * cz - sx * sy * sz), static_cast<float>(-sx * sy * cz - cx * sz), 0.0f,
-      0.0f);  // c3
-
-    h_ang.row(6).noalias() = Eigen::Vector4f(
-      static_cast<float>(-cy * cz), static_cast<float>(cy * sz), static_cast<float>(sy),
-      0.0f);  // d1
-    h_ang.row(7).noalias() = Eigen::Vector4f(
-      static_cast<float>(-sx * sy * cz), static_cast<float>(sx * sy * sz),
-      static_cast<float>(sx * cy), 0.0f);  // d2
-    h_ang.row(8).noalias() = Eigen::Vector4f(
-      static_cast<float>(cx * sy * cz), static_cast<float>(-cx * sy * sz),
-      static_cast<float>(-cx * cy), 0.0f);  // d3
-
-    h_ang.row(9).noalias() =
-      Eigen::Vector4f(static_cast<float>(sy * sz), static_cast<float>(sy * cz), 0.0f, 0.0f);  // e1
-    h_ang.row(10).noalias() = Eigen::Vector4f(
-      static_cast<float>(-sx * cy * sz), static_cast<float>(-sx * cy * cz), 0.0f, 0.0f);  // e2
-    h_ang.row(11).noalias() = Eigen::Vector4f(
-      static_cast<float>(cx * cy * sz), static_cast<float>(cx * cy * cz), 0.0f, 0.0f);  // e3
-
-    h_ang.row(12).noalias() =
-      Eigen::Vector4f(static_cast<float>(-cy * cz), static_cast<float>(cy * sz), 0.0f, 0.0f);  // f1
-    h_ang.row(13).noalias() = Eigen::Vector4f(
-      static_cast<float>(-cx * sz - sx * sy * cz), static_cast<float>(-cx * cz + sx * sy * sz),
-      0.0f, 0.0f);  // f2
-    h_ang.row(14).noalias() = Eigen::Vector4f(
-      static_cast<float>(-sx * sz + cx * sy * cz), static_cast<float>(-cx * sy * sz - sx * cz),
-      0.0f, 0.0f);  // f3
+    h_ang_.row(12) << (-cy * cz), (cy * sz), 0.0f, 0.0f;                                 // f1
+    h_ang_.row(13) << (-cx * sz - sx * sy * cz), (-cx * cz + sx * sy * sz), 0.0f, 0.0f;  // f2
+    h_ang_.row(14) << (-sx * sz + cx * sy * cz), (-cx * sy * sz - sx * cz), 0.0f, 0.0f;  // f3
   }
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computePointDerivatives(
-  Eigen::Vector3d & x, Eigen::Matrix<float, 4, 6> & point_gradient_,
-  Eigen::Matrix<float, 24, 6> & point_hessian_, bool compute_hessian) const
+void MultiGridNormalDistributionsTransform<PointSource, PointTarget>::computePointDerivatives(
+  Eigen::Vector3d & x, Eigen::Matrix<double, 4, 6> & point_gradient,
+  Eigen::Matrix<double, 24, 6> & point_hessian, bool compute_hessian) const
 {
-  Eigen::Vector4f x4(
-    static_cast<float>(x[0]), static_cast<float>(x[1]), static_cast<float>(x[2]), 0.0f);
+  Eigen::Vector4d x4(x[0], x[1], x[2], 0.0f);
 
   // Calculate first derivative of Transformation Equation 6.17 w.r.t. transform vector p.
   // Derivative w.r.t. ith element of transform vector corresponds to column i, Equation 6.18
   // and 6.19 [Magnusson 2009]
-  Eigen::Matrix<float, 8, 1> x_j_ang = j_ang * x4;
+  auto x_j_ang = j_ang_ * x4;
 
-  point_gradient_(1, 3) = x_j_ang[0];
-  point_gradient_(2, 3) = x_j_ang[1];
-  point_gradient_(0, 4) = x_j_ang[2];
-  point_gradient_(1, 4) = x_j_ang[3];
-  point_gradient_(2, 4) = x_j_ang[4];
-  point_gradient_(0, 5) = x_j_ang[5];
-  point_gradient_(1, 5) = x_j_ang[6];
-  point_gradient_(2, 5) = x_j_ang[7];
+  point_gradient(1, 3) = x_j_ang[0];
+  point_gradient(2, 3) = x_j_ang[1];
+  point_gradient(0, 4) = x_j_ang[2];
+  point_gradient(1, 4) = x_j_ang[3];
+  point_gradient(2, 4) = x_j_ang[4];
+  point_gradient(0, 5) = x_j_ang[5];
+  point_gradient(1, 5) = x_j_ang[6];
+  point_gradient(2, 5) = x_j_ang[7];
 
   if (compute_hessian) {
-    Eigen::Matrix<float, 16, 1> x_h_ang = h_ang * x4;
+    auto x_h_ang = h_ang_ * x4;
 
     // Vectors from Equation 6.21 [Magnusson 2009]
-    Eigen::Vector4f a(0, x_h_ang[0], x_h_ang[1], 0.0f);
-    Eigen::Vector4f b(0, x_h_ang[2], x_h_ang[3], 0.0f);
-    Eigen::Vector4f c(0, x_h_ang[4], x_h_ang[5], 0.0f);
-    Eigen::Vector4f d(x_h_ang[6], x_h_ang[7], x_h_ang[8], 0.0f);
-    Eigen::Vector4f e(x_h_ang[9], x_h_ang[10], x_h_ang[11], 0.0f);
-    Eigen::Vector4f f(x_h_ang[12], x_h_ang[13], x_h_ang[14], 0.0f);
+    Eigen::Vector4d a(0.0f, x_h_ang[0], x_h_ang[1], 0.0f);
+    Eigen::Vector4d b(0.0f, x_h_ang[2], x_h_ang[3], 0.0f);
+    Eigen::Vector4d c(0.0f, x_h_ang[4], x_h_ang[5], 0.0f);
+    Eigen::Vector4d d(x_h_ang[6], x_h_ang[7], x_h_ang[8], 0.0f);
+    Eigen::Vector4d e(x_h_ang[9], x_h_ang[10], x_h_ang[11], 0.0f);
+    Eigen::Vector4d f(x_h_ang[12], x_h_ang[13], x_h_ang[14], 0.0f);
 
     // Calculate second derivative of Transformation Equation 6.17 w.r.t. transform vector p.
     // Derivative w.r.t. ith and jth elements of transform vector corresponds to the 3x1 block
     // matrix starting at (3i,j), Equation 6.20 and 6.21 [Magnusson 2009]
-    point_hessian_.block<4, 1>((9 / 3) * 4, 3) = a;
-    point_hessian_.block<4, 1>((12 / 3) * 4, 3) = b;
-    point_hessian_.block<4, 1>((15 / 3) * 4, 3) = c;
-    point_hessian_.block<4, 1>((9 / 3) * 4, 4) = b;
-    point_hessian_.block<4, 1>((12 / 3) * 4, 4) = d;
-    point_hessian_.block<4, 1>((15 / 3) * 4, 4) = e;
-    point_hessian_.block<4, 1>((9 / 3) * 4, 5) = c;
-    point_hessian_.block<4, 1>((12 / 3) * 4, 5) = e;
-    point_hessian_.block<4, 1>((15 / 3) * 4, 5) = f;
+    point_hessian.block<4, 1>(12, 3) = a;
+    point_hessian.block<4, 1>(16, 3) = b;
+    point_hessian.block<4, 1>(20, 3) = c;
+    point_hessian.block<4, 1>(12, 4) = b;
+    point_hessian.block<4, 1>(16, 4) = d;
+    point_hessian.block<4, 1>(20, 4) = e;
+    point_hessian.block<4, 1>(12, 5) = c;
+    point_hessian.block<4, 1>(16, 5) = e;
+    point_hessian.block<4, 1>(20, 5) = f;
   }
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computePointDerivatives(
-  Eigen::Vector3d & x, Eigen::Matrix<double, 3, 6> & point_gradient_,
-  Eigen::Matrix<double, 18, 6> & point_hessian_, bool compute_hessian) const
-{
-  // Calculate first derivative of Transformation Equation 6.17 w.r.t. transform vector p.
-  // Derivative w.r.t. ith element of transform vector corresponds to column i, Equation 6.18
-  // and 6.19 [Magnusson 2009]
-  point_gradient_(1, 3) = x.dot(j_ang_a_);
-  point_gradient_(2, 3) = x.dot(j_ang_b_);
-  point_gradient_(0, 4) = x.dot(j_ang_c_);
-  point_gradient_(1, 4) = x.dot(j_ang_d_);
-  point_gradient_(2, 4) = x.dot(j_ang_e_);
-  point_gradient_(0, 5) = x.dot(j_ang_f_);
-  point_gradient_(1, 5) = x.dot(j_ang_g_);
-  point_gradient_(2, 5) = x.dot(j_ang_h_);
-
-  if (compute_hessian) {
-    // Vectors from Equation 6.21 [Magnusson 2009]
-    Eigen::Vector3d a;
-    Eigen::Vector3d b;
-    Eigen::Vector3d c;
-    Eigen::Vector3d d;
-    Eigen::Vector3d e;
-    Eigen::Vector3d f;
-
-    a << 0, x.dot(h_ang_a2_), x.dot(h_ang_a3_);
-    b << 0, x.dot(h_ang_b2_), x.dot(h_ang_b3_);
-    c << 0, x.dot(h_ang_c2_), x.dot(h_ang_c3_);
-    d << x.dot(h_ang_d1_), x.dot(h_ang_d2_), x.dot(h_ang_d3_);
-    e << x.dot(h_ang_e1_), x.dot(h_ang_e2_), x.dot(h_ang_e3_);
-    f << x.dot(h_ang_f1_), x.dot(h_ang_f2_), x.dot(h_ang_f3_);
-
-    // Calculate second derivative of Transformation Equation 6.17 w.r.t. transform vector p.
-    // Derivative w.r.t. ith and jth elements of transform vector corresponds to the 3x1 block
-    // matrix starting at (3i,j), Equation 6.20 and 6.21 [Magnusson 2009]
-    point_hessian_.block<3, 1>(9, 3) = a;
-    point_hessian_.block<3, 1>(12, 3) = b;
-    point_hessian_.block<3, 1>(15, 3) = c;
-    point_hessian_.block<3, 1>(9, 4) = b;
-    point_hessian_.block<3, 1>(12, 4) = d;
-    point_hessian_.block<3, 1>(15, 4) = e;
-    point_hessian_.block<3, 1>(9, 5) = c;
-    point_hessian_.block<3, 1>(12, 5) = e;
-    point_hessian_.block<3, 1>(15, 5) = f;
-  }
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-template <typename PointSource, typename PointTarget>
-double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::updateDerivatives(
+double MultiGridNormalDistributionsTransform<PointSource, PointTarget>::updateDerivatives(
   Eigen::Matrix<double, 6, 1> & score_gradient, Eigen::Matrix<double, 6, 6> & hessian,
-  const Eigen::Matrix<float, 4, 6> & point_gradient4,
-  const Eigen::Matrix<float, 24, 6> & point_hessian_, const Eigen::Vector3d & x_trans,
+  const Eigen::Matrix<double, 4, 6> & point_gradient,
+  const Eigen::Matrix<double, 24, 6> & point_hessian, const Eigen::Vector3d & x_trans,
   const Eigen::Matrix3d & c_inv, bool compute_hessian) const
 {
-  Eigen::Matrix<float, 1, 4> x_trans4(
-    static_cast<float>(x_trans[0]), static_cast<float>(x_trans[1]), static_cast<float>(x_trans[2]),
-    0.0f);
-  Eigen::Matrix4f c_inv4 = Eigen::Matrix4f::Zero();
-  c_inv4.topLeftCorner(3, 3) = c_inv.cast<float>();
+  Eigen::Matrix<double, 1, 4> x_trans4(x_trans[0], x_trans[1], x_trans[2], 0.0f);
+  Eigen::Matrix4d c_inv4 = Eigen::Matrix4d::Zero();
 
-  float gauss_d2 = gauss_d2_;
+  c_inv4.topLeftCorner(3, 3) = c_inv;
 
   // e^(-d_2/2 * (x_k - mu_k)^T Sigma_k^-1 (x_k - mu_k)) Equation 6.9 [Magnusson 2009]
-  float e_x_cov_x = exp(-gauss_d2 * x_trans4.dot(x_trans4 * c_inv4) * 0.5f);
+  double e_x_cov_x = exp(-gauss_d2_ * x_trans4.dot(x_trans4 * c_inv4) * 0.5f);
   // Calculate probability of transformed points existence, Equation 6.9 [Magnusson 2009]
-  float score_inc = -gauss_d1_ * e_x_cov_x;
+  double score_inc = -gauss_d1_ * e_x_cov_x;
 
-  e_x_cov_x = gauss_d2 * e_x_cov_x;
+  e_x_cov_x = gauss_d2_ * e_x_cov_x;
 
   // Error checking for invalid values.
-  if (e_x_cov_x > 1 || e_x_cov_x < 0 || e_x_cov_x != e_x_cov_x) return (0);
+  if (e_x_cov_x > 1 || e_x_cov_x < 0 || e_x_cov_x != e_x_cov_x) {
+    return 0;
+  }
 
   // Reusable portion of Equation 6.12 and 6.13 [Magnusson 2009]
   e_x_cov_x *= gauss_d1_;
 
-  Eigen::Matrix<float, 4, 6> c_inv4_x_point_gradient4 = c_inv4 * point_gradient4;
-  Eigen::Matrix<float, 6, 1> x_trans4_dot_c_inv4_x_point_gradient4 =
+  Eigen::Matrix<double, 4, 6> c_inv4_x_point_gradient4 = c_inv4 * point_gradient;
+  Eigen::Matrix<double, 6, 1> x_trans4_dot_c_inv4_x_point_gradient4 =
     x_trans4 * c_inv4_x_point_gradient4;
 
   score_gradient.noalias() += (e_x_cov_x * x_trans4_dot_c_inv4_x_point_gradient4).cast<double>();
 
   if (compute_hessian) {
-    Eigen::Matrix<float, 1, 4> x_trans4_x_c_inv4 = x_trans4 * c_inv4;
-    Eigen::Matrix<float, 6, 6> point_gradient4_colj_dot_c_inv4_x_point_gradient4_col_i =
-      point_gradient4.transpose() * c_inv4_x_point_gradient4;
-    Eigen::Matrix<float, 6, 1> x_trans4_dot_c_inv4_x_ext_point_hessian_4ij;
+    Eigen::Matrix<double, 1, 4> x_trans4_x_c_inv4 = x_trans4 * c_inv4;
+    Eigen::Matrix<double, 6, 6> point_gradient4_colj_dot_c_inv4_x_point_gradient4_col_i =
+      point_gradient.transpose() * c_inv4_x_point_gradient4;
+    Eigen::Matrix<double, 6, 1> x_trans4_dot_c_inv4_x_ext_point_hessian_4ij;
 
-    for (int i = 0; i < 6; i++) {
+    for (int i = 0; i < 6; ++i) {
       // Sigma_k^-1 d(T(x,p))/dpi, Reusable portion of Equation 6.12 and 6.13 [Magnusson 2009]
       // Update gradient, Equation 6.12 [Magnusson 2009]
       x_trans4_dot_c_inv4_x_ext_point_hessian_4ij.noalias() =
-        x_trans4_x_c_inv4 * point_hessian_.block<4, 6>(i * 4, 0);
+        x_trans4_x_c_inv4 * point_hessian.block<4, 6>(i * 4, 0);
 
       for (int j = 0; j < hessian.cols(); j++) {
         // Update hessian, Equation 6.13 [Magnusson 2009]
         hessian(i, j) +=
-          e_x_cov_x * (-gauss_d2 * x_trans4_dot_c_inv4_x_point_gradient4(i) *
+          e_x_cov_x * (-gauss_d2_ * x_trans4_dot_c_inv4_x_point_gradient4(i) *
                          x_trans4_dot_c_inv4_x_point_gradient4(j) +
                        x_trans4_dot_c_inv4_x_ext_point_hessian_4ij(j) +
                        point_gradient4_colj_dot_c_inv4_x_point_gradient4_col_i(j, i));
@@ -693,32 +677,27 @@ double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::updateDer
     }
   }
 
-  return (score_inc);
+  return score_inc;
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeHessian(
+void MultiGridNormalDistributionsTransform<PointSource, PointTarget>::computeHessian(
   Eigen::Matrix<double, 6, 6> & hessian, PointCloudSource & trans_cloud,
   Eigen::Matrix<double, 6, 1> &)
 {
-  // Original Point and Transformed Point
-  PointSource x_pt;
-  PointSource x_trans_pt;
-  // Original Point and Transformed Point (for math)
-  Eigen::Vector3d x;
-  Eigen::Vector3d x_trans;
-  // Occupied Voxel
-  TargetGridLeafConstPtr cell;
-  // Inverse Covariance of Occupied Voxel
-  Eigen::Matrix3d c_inv;
-
   // Initialize Point Gradient and Hessian
-  Eigen::Matrix<double, 3, 6> point_gradient;
-  Eigen::Matrix<double, 18, 6> point_hessian;
-  point_gradient.setZero();
-  point_gradient.block<3, 3>(0, 0).setIdentity();
-  point_hessian.setZero();
+  // Pre-allocate thread-wise point gradients and point hessians
+  std::vector<Eigen::Matrix<double, 4, 6>> t_point_gradients(params_.num_threads);
+  std::vector<Eigen::Matrix<double, 24, 6>> t_point_hessians(params_.num_threads);
+  std::vector<Eigen::Matrix<double, 6, 6>> t_hessians(params_.num_threads);
+
+  for (int i = 0; i < params_.num_threads; ++i) {
+    t_point_gradients[i].setZero();
+    t_point_gradients[i].block<3, 3>(0, 0).setIdentity();
+    t_point_hessians[i].setZero();
+    t_hessians[i].setZero();
+  }
 
   hessian.setZero();
 
@@ -726,54 +705,54 @@ void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeHess
   // calculation
 
   // Update hessian for each point, line 17 in Algorithm 2 [Magnusson 2009]
-  for (size_t idx = 0; idx < input_->points.size(); idx++) {
-    x_trans_pt = trans_cloud.points[idx];
+#pragma omp parallel for num_threads(params_.num_threads) schedule(guided, 8)
+  for (size_t idx = 0; idx < input_->size(); ++idx) {
+    int tid = omp_get_thread_num();
+    auto & x_trans_pt = trans_cloud[idx];
 
     // Find neighbors (Radius search has been experimentally faster than direct neighbor checking.
     std::vector<TargetGridLeafConstPtr> neighborhood;
-    std::vector<float> distances;
-    switch (params_.search_method) {
-      case KDTREE:
-        target_cells_.radiusSearch(x_trans_pt, params_.resolution, neighborhood, distances);
-        break;
-      case DIRECT26:
-        target_cells_.getNeighborhoodAtPoint(x_trans_pt, neighborhood);
-        break;
-      default:
-      case DIRECT7:
-        target_cells_.getNeighborhoodAtPoint7(x_trans_pt, neighborhood);
-        break;
-      case DIRECT1:
-        target_cells_.getNeighborhoodAtPoint1(x_trans_pt, neighborhood);
-        break;
+
+    // Neighborhood search method other than kdtree is disabled in multigrid_ndt_omp
+    target_cells_.radiusSearch(x_trans_pt, params_.resolution, neighborhood);
+
+    if (neighborhood.empty()) {
+      continue;
     }
 
-    for (auto & cell : neighborhood) {
-      x_pt = input_->points[idx];
-      x = Eigen::Vector3d(x_pt.x, x_pt.y, x_pt.z);
+    auto & x_pt = (*input_)[idx];
+    // For math
+    Eigen::Vector3d x(x_pt.x, x_pt.y, x_pt.z);
+    const Eigen::Vector3d x_trans(x_trans_pt.x, x_trans_pt.y, x_trans_pt.z);
 
-      x_trans = Eigen::Vector3d(x_trans_pt.x, x_trans_pt.y, x_trans_pt.z);
+    auto & point_gradient = t_point_gradients[tid];
+    auto & point_hessian = t_point_hessians[tid];
+    auto & tmp_hessian = t_hessians[tid];
 
-      // Denorm point, x_k' in Equations 6.12 and 6.13 [Magnusson 2009]
-      x_trans -= cell->getMean();
-      // Uses precomputed covariance for speed.
-      c_inv = cell->getInverseCov();
+    // Compute derivative of transform function w.r.t. transform vector, J_E and H_E in
+    // Equations 6.18 and 6.20 [Magnusson 2009]
+    computePointDerivatives(x, point_gradient, point_hessian);
 
-      // Compute derivative of transform function w.r.t. transform vector, J_E and H_E in
-      // Equations 6.18 and 6.20 [Magnusson 2009]
-      computePointDerivatives(x, point_gradient, point_hessian);
+    for (const auto & cell : neighborhood) {
       // Update hessian, lines 21 in Algorithm 2, according to Equations 6.10, 6.12 and 6.13,
       // respectively [Magnusson 2009]
-      updateHessian(hessian, point_gradient, point_hessian, x_trans, c_inv);
+      updateHessian(
+        tmp_hessian, point_gradient, point_hessian, x_trans - cell->getMean(),
+        cell->getInverseCov());
     }
+  }
+
+  // Sum over t_hessians
+  for (const auto & th : t_hessians) {
+    hessian += th;
   }
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::updateHessian(
-  Eigen::Matrix<double, 6, 6> & hessian, const Eigen::Matrix<double, 3, 6> & point_gradient_,
-  const Eigen::Matrix<double, 18, 6> & point_hessian_, const Eigen::Vector3d & x_trans,
+void MultiGridNormalDistributionsTransform<PointSource, PointTarget>::updateHessian(
+  Eigen::Matrix<double, 6, 6> & hessian, const Eigen::Matrix<double, 4, 6> & point_gradient,
+  const Eigen::Matrix<double, 24, 6> & point_hessian, const Eigen::Vector3d & x_trans,
   const Eigen::Matrix3d & c_inv) const
 {
   Eigen::Vector3d cov_dxd_pi;
@@ -781,28 +760,32 @@ void pclomp::NormalDistributionsTransform<PointSource, PointTarget>::updateHessi
   double e_x_cov_x = gauss_d2_ * exp(-gauss_d2_ * x_trans.dot(c_inv * x_trans) / 2);
 
   // Error checking for invalid values.
-  if (e_x_cov_x > 1 || e_x_cov_x < 0 || e_x_cov_x != e_x_cov_x) return;
+  if (e_x_cov_x > 1 || e_x_cov_x < 0 || e_x_cov_x != e_x_cov_x) {
+    return;
+  }
 
   // Reusable portion of Equation 6.12 and 6.13 [Magnusson 2009]
   e_x_cov_x *= gauss_d1_;
 
-  for (int i = 0; i < 6; i++) {
+  for (int i = 0; i < 6; ++i) {
     // Sigma_k^-1 d(T(x,p))/dpi, Reusable portion of Equation 6.12 and 6.13 [Magnusson 2009]
-    cov_dxd_pi = c_inv * point_gradient_.col(i);
+    cov_dxd_pi = c_inv * point_gradient.block<3, 1>(0, i);
 
     for (int j = 0; j < hessian.cols(); j++) {
       // Update hessian, Equation 6.13 [Magnusson 2009]
-      hessian(i, j) += e_x_cov_x * (-gauss_d2_ * x_trans.dot(cov_dxd_pi) *
-                                      x_trans.dot(c_inv * point_gradient_.col(j)) +
-                                    x_trans.dot(c_inv * point_hessian_.block<3, 1>(3 * i, j)) +
-                                    point_gradient_.col(j).dot(cov_dxd_pi));
+      Eigen::Vector3d pg_col = point_gradient.block<3, 1>(0, j);
+
+      hessian(i, j) +=
+        e_x_cov_x *
+        (-gauss_d2_ * x_trans.dot(cov_dxd_pi) * x_trans.dot(c_inv * pg_col) +
+         x_trans.dot(c_inv * point_hessian.block<3, 1>(3 * i, j)) + pg_col.dot(cov_dxd_pi));
     }
   }
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-bool pclomp::NormalDistributionsTransform<PointSource, PointTarget>::updateIntervalMT(
+bool MultiGridNormalDistributionsTransform<PointSource, PointTarget>::updateIntervalMT(
   double & a_l, double & f_l, double & g_l, double & a_u, double & f_u, double & g_u, double a_t,
   double f_t, double g_t)
 {
@@ -811,14 +794,14 @@ bool pclomp::NormalDistributionsTransform<PointSource, PointTarget>::updateInter
     a_u = a_t;
     f_u = f_t;
     g_u = g_t;
-    return (false);
+    return false;
   }
   // Case U2 in Update Algorithm and Case b in Modified Update Algorithm [More, Thuente 1994]
   if (g_t * (a_l - a_t) > 0) {
     a_l = a_t;
     f_l = f_t;
     g_l = g_t;
-    return (false);
+    return false;
   }
   // Case U3 in Update Algorithm and Case c in Modified Update Algorithm [More, Thuente 1994]
   if (g_t * (a_l - a_t) < 0) {
@@ -829,15 +812,15 @@ bool pclomp::NormalDistributionsTransform<PointSource, PointTarget>::updateInter
     a_l = a_t;
     f_l = f_t;
     g_l = g_t;
-    return (false);
+    return false;
   }
   // Interval Converged
-  return (true);
+  return true;
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::trialValueSelectionMT(
+double MultiGridNormalDistributionsTransform<PointSource, PointTarget>::trialValueSelectionMT(
   double a_l, double f_l, double g_l, double a_u, double f_u, double g_u, double a_t, double f_t,
   double g_t)
 {
@@ -855,9 +838,9 @@ double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::trialValu
     double a_q = a_l - 0.5 * (a_l - a_t) * g_l / (g_l - (f_l - f_t) / (a_l - a_t));
 
     if (std::fabs(a_c - a_l) < std::fabs(a_q - a_l)) {
-      return (a_c);
+      return a_c;
     }
-    return (0.5 * (a_q + a_c));
+    return 0.5 * (a_q + a_c);
   }
   // Case 2 in Trial Value Selection [More, Thuente 1994]
   if (g_t * g_l < 0) {
@@ -873,9 +856,9 @@ double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::trialValu
     double a_s = a_l - (a_l - a_t) / (g_l - g_t) * g_l;
 
     if (std::fabs(a_c - a_t) >= std::fabs(a_s - a_t)) {
-      return (a_c);
+      return a_c;
     }
-    return (a_s);
+    return a_s;
   }
   // Case 3 in Trial Value Selection [More, Thuente 1994]
   if (std::fabs(g_t) <= std::fabs(g_l)) {
@@ -889,17 +872,18 @@ double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::trialValu
     // Equation 2.4.5 [Sun, Yuan 2006]
     double a_s = a_l - (a_l - a_t) / (g_l - g_t) * g_l;
 
-    double a_t_next = NAN;
+    double a_t_next = 0.0;
 
-    if (std::fabs(a_c - a_t) < std::fabs(a_s - a_t))
+    if (std::fabs(a_c - a_t) < std::fabs(a_s - a_t)) {
       a_t_next = a_c;
-    else
+    } else {
       a_t_next = a_s;
+    }
 
     if (a_t > a_l) {
-      return (std::min(a_t + 0.66 * (a_u - a_t), a_t_next));
+      return std::min(a_t + 0.66 * (a_u - a_t), a_t_next);
     }
-    return (std::max(a_t + 0.66 * (a_u - a_t), a_t_next));
+    return std::max(a_t + 0.66 * (a_u - a_t), a_t_next);
   }
   // Case 4 in Trial Value Selection [More, Thuente 1994]
   // Calculate the minimizer of the cubic that interpolates f_u, f_t, g_u and g_t
@@ -907,12 +891,12 @@ double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::trialValu
   double z = 3 * (f_t - f_u) / (a_t - a_u) - g_t - g_u;
   double w = std::sqrt(z * z - g_t * g_u);
   // Equation 2.4.56 [Sun, Yuan 2006]
-  return (a_u + (a_t - a_u) * (w - g_u - z) / (g_t - g_u + 2 * w));
+  return a_u + (a_t - a_u) * (w - g_u - z) / (g_t - g_u + 2 * w);
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointSource, typename PointTarget>
-double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeStepLengthMT(
+double MultiGridNormalDistributionsTransform<PointSource, PointTarget>::computeStepLengthMT(
   const Eigen::Matrix<double, 6, 1> & x, Eigen::Matrix<double, 6, 1> & step_dir, double step_init,
   double step_max, double step_min, double & score, Eigen::Matrix<double, 6, 1> & score_gradient,
   Eigen::Matrix<double, 6, 6> & hessian, PointCloudSource & trans_cloud)
@@ -926,7 +910,9 @@ double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeSt
 
   if (d_phi_0 >= 0) {
     // Not a decent direction
-    if (d_phi_0 == 0) return 0;
+    if (d_phi_0 == 0) {
+      return 0;
+    }
 
     // Reverse step direction and calculate optimal step.
     d_phi_0 *= -1;
@@ -1073,192 +1059,107 @@ double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::computeSt
   // If inner loop was run then hessian needs to be calculated.
   // Hessian is unnecessary for step length determination but gradients are required
   // so derivative and transform data is stored for the next iteration.
-  if (step_iterations) computeHessian(hessian, trans_cloud, x_t);
-
-  return (a_t);
-}
-
-// change at 20220721 konishi
-template <typename PointSource, typename PointTarget>
-double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::calculateScore(
-  const PointCloudSource & trans_cloud)
-{
-  double score = 0;
-  std::map<size_t, size_t> voxel_points_num;
-
-  for (std::size_t idx = 0; idx < trans_cloud.points.size(); idx++) {
-    PointSource x_trans_pt = trans_cloud.points[idx];
-
-    // Find neighbors (Radius search has been experimentally faster than direct neighbor checking.
-    std::vector<TargetGridLeafConstPtr> neighborhood;
-    std::vector<float> distances;
-    switch (params_.search_method) {
-      case KDTREE:
-        target_cells_.radiusSearch(x_trans_pt, params_.resolution, neighborhood, distances);
-        break;
-      case DIRECT26:
-        target_cells_.getNeighborhoodAtPoint(x_trans_pt, neighborhood);
-        break;
-      default:
-      case DIRECT7:
-        target_cells_.getNeighborhoodAtPoint7(x_trans_pt, neighborhood);
-        break;
-      case DIRECT1:
-        target_cells_.getNeighborhoodAtPoint1(x_trans_pt, neighborhood);
-        break;
-    }
-
-    // add at 20220218 by konishi
-    size_t voxel_idx;
-
-    if (neighborhood.size() == 0) {
-      // Compute the 3D index of the voxel containing the query point
-      Eigen::Vector3i vid;
-
-      vid(0) = static_cast<int>(floor(x_trans_pt.x / params_.resolution));
-      vid(1) = static_cast<int>(floor(x_trans_pt.y / params_.resolution));
-      vid(2) = static_cast<int>(floor(x_trans_pt.z / params_.resolution));
-
-      empty_voxels_.insert(vid);
-
-      if (nomap_points_num_.count(voxel_idx) == 0) {
-        nomap_points_num_[voxel_idx] = 0;
-      }
-      nomap_points_num_[voxel_idx] += 1;
-
-    } else {
-      for (auto & cell : neighborhood) {
-        PointSource x_pt = input_->points[idx];
-        Eigen::Vector3d x = Eigen::Vector3d(x_pt.x, x_pt.y, x_pt.z);
-
-        Eigen::Vector3d x_trans = Eigen::Vector3d(x_trans_pt.x, x_trans_pt.y, x_trans_pt.z);
-
-        // Denorm point, x_k' in Equations 6.12 and 6.13 [Magnusson 2009]
-        x_trans -= cell->getMean();
-        // Uses precomputed covariance for speed.
-        Eigen::Matrix3d c_inv = cell->getInverseCov();
-
-        // e^(-d_2/2 * (x_k - mu_k)^T Sigma_k^-1 (x_k - mu_k)) Equation 6.9 [Magnusson 2009]
-        double e_x_cov_x = exp(-gauss_d2_ * x_trans.dot(c_inv * x_trans) / 2);
-        // Calculate probability of transformed points existence, Equation 6.9 [Magnusson 2009]
-        double score_inc = -gauss_d1_ * e_x_cov_x;
-
-        score += score_inc;
-
-        voxel_idx = target_cells_.getLeafIndex(cell->getMean());
-
-        if (voxel_points_num.count(voxel_idx) == 0) {
-          voxel_points_num[voxel_idx] = 0;
-          voxel_score_map_[voxel_idx] = 0;
-        }
-
-        voxel_score_map_[voxel_idx] += score_inc;
-        voxel_points_num[voxel_idx] += 1;
-      }
-    }
-  }
-  for (auto & voxel_score_output : voxel_score_map_) {
-    if (voxel_points_num[voxel_score_output.first] != 0) {
-      voxel_score_output.second /= (voxel_points_num[voxel_score_output.first]);
-    }
+  if (step_iterations) {
+    computeHessian(hessian, trans_cloud, x_t);
   }
 
-  return (score) / static_cast<double>(trans_cloud.size());
+  return a_t;
 }
 
 template <typename PointSource, typename PointTarget>
 double
-pclomp::NormalDistributionsTransform<PointSource, PointTarget>::calculateTransformationProbability(
+MultiGridNormalDistributionsTransform<PointSource, PointTarget>::calculateTransformationProbability(
   const PointCloudSource & trans_cloud) const
 {
   double score = 0;
 
-  for (std::size_t idx = 0; idx < trans_cloud.points.size(); idx++) {
-    PointSource x_trans_pt = trans_cloud.points[idx];
+  // Score per thread
+  std::vector<double> t_scores(params_.num_threads, 0);
+
+#pragma omp parallel for num_threads(params_.num_threads) schedule(guided, 8)
+  for (size_t idx = 0; idx < trans_cloud.size(); ++idx) {
+    int tid = omp_get_thread_num();
+    PointSource x_trans_pt = trans_cloud[idx];
 
     // Find neighbors (Radius search has been experimentally faster than direct neighbor checking.
     std::vector<TargetGridLeafConstPtr> neighborhood;
-    std::vector<float> distances;
-    switch (params_.search_method) {
-      case KDTREE:
-        target_cells_.radiusSearch(x_trans_pt, params_.resolution, neighborhood, distances);
-        break;
-      case DIRECT26:
-        target_cells_.getNeighborhoodAtPoint(x_trans_pt, neighborhood);
-        break;
-      default:
-      case DIRECT7:
-        target_cells_.getNeighborhoodAtPoint7(x_trans_pt, neighborhood);
-        break;
-      case DIRECT1:
-        target_cells_.getNeighborhoodAtPoint1(x_trans_pt, neighborhood);
-        break;
+
+    // Neighborhood search method other than kdtree is disabled in multigrid_ndt_omp
+    target_cells_.radiusSearch(x_trans_pt, params_.resolution, neighborhood);
+
+    if (neighborhood.empty()) {
+      continue;
     }
 
+    double tmp_score = 0;
+
     for (auto & cell : neighborhood) {
-      Eigen::Vector3d x_trans = Eigen::Vector3d(x_trans_pt.x, x_trans_pt.y, x_trans_pt.z);
+      Eigen::Vector3d x_trans(x_trans_pt.x, x_trans_pt.y, x_trans_pt.z);
 
       // Denorm point, x_k' in Equations 6.12 and 6.13 [Magnusson 2009]
       x_trans -= cell->getMean();
-      // Uses precomputed covariance for speed.
-      Eigen::Matrix3d c_inv = cell->getInverseCov();
 
-      // e^(-d_2/2 * (x_k - mu_k)^T Sigma_k^-1 (x_k - mu_k)) Equation 6.9 [Magnusson 2009]
-      double e_x_cov_x = exp(-gauss_d2_ * x_trans.dot(c_inv * x_trans) / 2);
       // Calculate probability of transformed points existence, Equation 6.9 [Magnusson 2009]
-      double score_inc = -gauss_d1_ * e_x_cov_x;
-
-      score += score_inc;
+      // e^(-d_2/2 * (x_k - mu_k)^T Sigma_k^-1 (x_k - mu_k)) Equation 6.9 [Magnusson 2009]
+      tmp_score -= gauss_d1_ * exp(-gauss_d2_ * x_trans.dot(cell->getInverseCov() * x_trans) / 2.0);
     }
+
+    t_scores[tid] += tmp_score;
+  }
+
+  // Sum the point-wise scores
+  for (const auto & ts : t_scores) {
+    score += ts;
   }
 
   double output_score = 0;
-  if (!trans_cloud.points.empty()) {
-    output_score = (score) / static_cast<double>(trans_cloud.points.size());
+  if (!trans_cloud.empty()) {
+    output_score = (score) / static_cast<double>(trans_cloud.size());
   }
   return output_score;
 }
 
 template <typename PointSource, typename PointTarget>
-double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::
+double MultiGridNormalDistributionsTransform<PointSource, PointTarget>::
   calculateNearestVoxelTransformationLikelihood(const PointCloudSource & trans_cloud) const
 {
   double nearest_voxel_score = 0;
-  size_t found_neigborhood_voxel_num = 0;
+  size_t found_neighborhood_voxel_num = 0;
 
-  for (std::size_t idx = 0; idx < trans_cloud.points.size(); idx++) {
-    double nearest_voxel_score_pt = 0;
-    PointSource x_trans_pt = trans_cloud.points[idx];
+  // Thread-wise results
+  std::vector<double> t_nvs(params_.num_threads);
+  std::vector<size_t> t_found_nnvn(params_.num_threads);
+
+  for (int i = 0; i < params_.num_threads; ++i) {
+    t_nvs[i] = 0;
+    t_found_nnvn[i] = 0;
+  }
+
+#pragma omp parallel for num_threads(params_.num_threads) schedule(guided, 8)
+  for (size_t idx = 0; idx < trans_cloud.size(); ++idx) {
+    int tid = omp_get_thread_num();
+    PointSource x_trans_pt = trans_cloud[idx];
 
     // Find neighbors (Radius search has been experimentally faster than direct neighbor checking.
     std::vector<TargetGridLeafConstPtr> neighborhood;
-    std::vector<float> distances;
-    switch (params_.search_method) {
-      case KDTREE:
-        target_cells_.radiusSearch(x_trans_pt, params_.resolution, neighborhood, distances);
-        break;
-      case DIRECT26:
-        target_cells_.getNeighborhoodAtPoint(x_trans_pt, neighborhood);
-        break;
-      default:
-      case DIRECT7:
-        target_cells_.getNeighborhoodAtPoint7(x_trans_pt, neighborhood);
-        break;
-      case DIRECT1:
-        target_cells_.getNeighborhoodAtPoint1(x_trans_pt, neighborhood);
-        break;
+
+    // Neighborhood search method other than kdtree is disabled in multigrid_ndt_omp
+    target_cells_.radiusSearch(x_trans_pt, params_.resolution, neighborhood);
+
+    if (neighborhood.empty()) {
+      continue;
     }
 
+    double nearest_voxel_score_pt = 0;
+
     for (auto & cell : neighborhood) {
-      Eigen::Vector3d x_trans = Eigen::Vector3d(x_trans_pt.x, x_trans_pt.y, x_trans_pt.z);
+      Eigen::Vector3d x_trans(x_trans_pt.x, x_trans_pt.y, x_trans_pt.z);
 
       // Denorm point, x_k' in Equations 6.12 and 6.13 [Magnusson 2009]
       x_trans -= cell->getMean();
-      // Uses precomputed covariance for speed.
-      Eigen::Matrix3d c_inv = cell->getInverseCov();
 
       // e^(-d_2/2 * (x_k - mu_k)^T Sigma_k^-1 (x_k - mu_k)) Equation 6.9 [Magnusson 2009]
-      double e_x_cov_x = exp(-gauss_d2_ * x_trans.dot(c_inv * x_trans) / 2);
+      double e_x_cov_x = exp(-gauss_d2_ * x_trans.dot(cell->getInverseCov() * x_trans) / 2.0);
       // Calculate probability of transformed points existence, Equation 6.9 [Magnusson 2009]
       double score_inc = -gauss_d1_ * e_x_cov_x;
 
@@ -1267,17 +1168,83 @@ double pclomp::NormalDistributionsTransform<PointSource, PointTarget>::
       }
     }
 
-    if (!neighborhood.empty()) {
-      ++found_neigborhood_voxel_num;
-      nearest_voxel_score += nearest_voxel_score_pt;
-    }
+    t_nvs[tid] += nearest_voxel_score_pt;
+    ++t_found_nnvn[tid];
+  }
+
+  // Sum up point-wise scores
+  for (size_t idx = 0; idx < params_.num_threads; ++idx) {
+    found_neighborhood_voxel_num += t_found_nnvn[idx];
+    nearest_voxel_score += t_nvs[idx];
   }
 
   double output_score = 0;
-  if (found_neigborhood_voxel_num != 0) {
-    output_score = nearest_voxel_score / static_cast<double>(found_neigborhood_voxel_num);
+
+  if (found_neighborhood_voxel_num != 0) {
+    output_score = nearest_voxel_score / static_cast<double>(found_neighborhood_voxel_num);
   }
   return output_score;
 }
 
-#endif  // PCL_REGISTRATION_NDT_IMPL_H_
+template <typename PointSource, typename PointTarget>
+pcl::PointCloud<pcl::PointXYZI> MultiGridNormalDistributionsTransform<PointSource, PointTarget>::
+  calculateNearestVoxelScoreEachPoint(const PointCloudSource & trans_cloud) const
+{
+  // Thread-wise results
+  std::vector<pcl::PointCloud<pcl::PointXYZI>> threads_pc(params_.num_threads);
+  pcl::PointCloud<pcl::PointXYZI> score_points;
+
+#pragma omp parallel for num_threads(params_.num_threads) schedule(guided, 8)
+  for (size_t idx = 0; idx < trans_cloud.size(); ++idx) {
+    int tid = omp_get_thread_num();
+    PointSource x_trans_pt = trans_cloud[idx];
+
+    // Find neighbors (Radius search has been experimentally faster than direct neighbor checking.
+    std::vector<TargetGridLeafConstPtr> neighborhood;
+
+    // Neighborhood search method other than kdtree is disabled in multigrid_ndt_omp
+    target_cells_.radiusSearch(x_trans_pt, params_.resolution, neighborhood);
+
+    if (neighborhood.empty()) {
+      continue;
+    }
+
+    double nearest_voxel_score_pt = 0;
+
+    for (auto & cell : neighborhood) {
+      Eigen::Vector3d x_trans(x_trans_pt.x, x_trans_pt.y, x_trans_pt.z);
+
+      // Denorm point, x_k' in Equations 6.12 and 6.13 [Magnusson 2009]
+      x_trans -= cell->getMean();
+
+      // e^(-d_2/2 * (x_k - mu_k)^T Sigma_k^-1 (x_k - mu_k)) Equation 6.9 [Magnusson 2009]
+      double e_x_cov_x = exp(-gauss_d2_ * x_trans.dot(cell->getInverseCov() * x_trans) / 2.0);
+      // Calculate probability of transformed points existence, Equation 6.9 [Magnusson 2009]
+      double score_inc = -gauss_d1_ * e_x_cov_x;
+
+      if (score_inc > nearest_voxel_score_pt) {
+        nearest_voxel_score_pt = score_inc;
+      }
+    }
+
+    pcl::PointXYZI sensor_point_score;
+    sensor_point_score.x = trans_cloud.points[idx].x;
+    sensor_point_score.y = trans_cloud.points[idx].y;
+    sensor_point_score.z = trans_cloud.points[idx].z;
+    sensor_point_score.intensity = nearest_voxel_score_pt;
+    threads_pc[tid].points.push_back(sensor_point_score);
+  }
+
+  // Sum up point-wise scores
+  for (size_t idx = 0; idx < params_.num_threads; ++idx) {
+    for (size_t p_idx = 0; p_idx < threads_pc[idx].size(); ++p_idx) {
+      score_points.points.push_back(threads_pc[idx].points[p_idx]);
+    }
+  }
+
+  return score_points;
+}
+
+}  // namespace autoware::ndt_omp::pclomp
+
+#endif  // PCL_REGISTRATION_NDT_OMP_MULTI_VOXEL_IMPL_H_
