+++ include/autoware/ndt_omp/pclomp/voxel_grid_covariance_omp_impl.hpp
@@ -52,7 +52,7 @@
 
 //////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointT>
-void pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
+void autoware::ndt_omp::pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
 {
   voxel_centroids_leaf_indices_.clear();
 
@@ -72,12 +72,13 @@ void pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
   Eigen::Vector4f min_p;
   Eigen::Vector4f max_p;
   // Get the minimum and maximum dimensions
-  if (!filter_field_name_.empty())  // If we don't want to process the entire cloud...
+  if (!filter_field_name_.empty()) {  // If we don't want to process the entire cloud...
     pcl::getMinMax3D<PointT>(
       input_, filter_field_name_, static_cast<float>(filter_limit_min_),
       static_cast<float>(filter_limit_max_), min_p, max_p, filter_limit_negative_);
-  else
+  } else {
     pcl::getMinMax3D<PointT>(*input_, min_p, max_p);
+  }
 
   // Check that the leaf size is not too small, given the size of the data
   int64_t dx = static_cast<int64_t>((max_p[0] - min_p[0]) * inverse_leaf_size_[0]) + 1;
@@ -114,13 +115,17 @@ void pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
 
   int centroid_size = 4;
 
-  if (downsample_all_data_) centroid_size = boost::mpl::size<FieldList>::value;
+  if (downsample_all_data_) {
+    centroid_size = boost::mpl::size<FieldList>::value;
+  }
 
   // ---[ RGB special case
   std::vector<pcl::PCLPointField> fields;
   int rgba_index = -1;
   rgba_index = pcl::getFieldIndex<PointT>("rgb", fields);
-  if (rgba_index == -1) rgba_index = pcl::getFieldIndex<PointT>("rgba", fields);
+  if (rgba_index == -1) {
+    rgba_index = pcl::getFieldIndex<PointT>("rgba", fields);
+  }
   if (rgba_index >= 0) {
     rgba_index = static_cast<int>(fields[rgba_index].offset);
     centroid_size += 3;
@@ -132,19 +137,22 @@ void pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
     // Get the distance field index
     std::vector<pcl::PCLPointField> distance_fields;
     int distance_idx = pcl::getFieldIndex<PointT>(filter_field_name_, distance_fields);
-    if (distance_idx == -1)
+    if (distance_idx == -1) {
       PCL_WARN(
         "[pcl::%s::applyFilter] Invalid filter field name. Index is %d.\n", getClassName().c_str(),
         distance_idx);
+    }
 
     // First pass: go over all points and insert them into the right leaf
     for (size_t cp = 0; cp < input_->points.size(); ++cp) {
-      if (!input_->is_dense)
+      if (!input_->is_dense) {
         // Check if the point is invalid
         if (
           !std::isfinite(input_->points[cp].x) || !std::isfinite(input_->points[cp].y) ||
-          !std::isfinite(input_->points[cp].z))
+          !std::isfinite(input_->points[cp].z)) {
           continue;
+        }
+      }
 
       // Get the distance value
       const auto * pt_data = reinterpret_cast<const uint8_t *>(&input_->points[cp]);
@@ -153,10 +161,14 @@ void pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
 
       if (filter_limit_negative_) {
         // Use a threshold for cutting out points which inside the interval
-        if ((distance_value < filter_limit_max_) && (distance_value > filter_limit_min_)) continue;
+        if ((distance_value < filter_limit_max_) && (distance_value > filter_limit_min_)) {
+          continue;
+        }
       } else {
         // Use a threshold for cutting out points which are too close/far away
-        if ((distance_value > filter_limit_max_) || (distance_value < filter_limit_min_)) continue;
+        if ((distance_value > filter_limit_max_) || (distance_value < filter_limit_min_)) {
+          continue;
+        }
       }
 
       int ijk0 = static_cast<int>(
@@ -209,12 +221,14 @@ void pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
   } else {
     // First pass: go over all points and insert them into the right leaf
     for (size_t cp = 0; cp < input_->points.size(); ++cp) {
-      if (!input_->is_dense)
+      if (!input_->is_dense) {
         // Check if the point is invalid
         if (
           !std::isfinite(input_->points[cp].x) || !std::isfinite(input_->points[cp].y) ||
-          !std::isfinite(input_->points[cp].z))
+          !std::isfinite(input_->points[cp].z)) {
           continue;
+        }
+      }
 
       int ijk0 = static_cast<int>(
         floor(input_->points[cp].x * inverse_leaf_size_[0]) - static_cast<float>(min_b_[0]));
@@ -265,9 +279,13 @@ void pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
 
   // Second pass: go over all leaves and compute centroids and covariance matrices
   output.points.reserve(leaves_.size());
-  if (searchable_) voxel_centroids_leaf_indices_.reserve(leaves_.size());
+  if (searchable_) {
+    voxel_centroids_leaf_indices_.reserve(leaves_.size());
+  }
   int cp = 0;
-  if (save_leaf_layout_) leaf_layout_.resize(div_b_[0] * div_b_[1] * div_b_[2], -1);
+  if (save_leaf_layout_) {
+    leaf_layout_.resize(div_b_[0] * div_b_[1] * div_b_[2], -1);
+  }
 
   // Eigen values and vectors calculated to prevent near singular matrices
   Eigen::SelfAdjointEigenSolver<Eigen::Matrix3d> eigensolver;
@@ -293,7 +311,9 @@ void pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
     // voxel centroids and output clouds. Points with less than the minimum points will have a can
     // not be accurately approximated using a normal distribution.
     if (leaf.nr_points_ >= min_points_per_voxel_) {
-      if (save_leaf_layout_) leaf_layout_[it->first] = cp++;
+      if (save_leaf_layout_) {
+        leaf_layout_[it->first] = cp++;
+      }
 
       output.push_back(PointT());
 
@@ -320,7 +340,9 @@ void pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
       }
 
       // Stores the voxel indices for fast access searching
-      if (searchable_) voxel_centroids_leaf_indices_.push_back(static_cast<int>(it->first));
+      if (searchable_) {
+        voxel_centroids_leaf_indices_.push_back(static_cast<int>(it->first));
+      }
 
       // Single pass covariance calculation
       leaf.cov_ = (leaf.cov_ - 2 * (pt_sum * leaf.mean_.transpose())) / leaf.nr_points_ +
@@ -365,7 +387,7 @@ void pclomp::VoxelGridCovariance<PointT>::applyFilter(PointCloud & output)
 
 //////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointT>
-int pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint(
+int autoware::ndt_omp::pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint(
   const Eigen::MatrixXi & relative_coordinates, const PointT & reference_point,
   std::vector<LeafConstPtr> & neighbors) const
 {
@@ -395,12 +417,12 @@ int pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint(
     }
   }
 
-  return (static_cast<int>(neighbors.size()));
+  return static_cast<int>(neighbors.size());
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointT>
-int pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint(
+int autoware::ndt_omp::pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint(
   const PointT & reference_point, std::vector<LeafConstPtr> & neighbors) const
 {
   neighbors.clear();
@@ -412,7 +434,7 @@ int pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint(
 
 //////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointT>
-int pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint7(
+int autoware::ndt_omp::pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint7(
   const PointT & reference_point, std::vector<LeafConstPtr> & neighbors) const
 {
   neighbors.clear();
@@ -431,7 +453,7 @@ int pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint7(
 
 //////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointT>
-int pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint1(
+int autoware::ndt_omp::pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint1(
   const PointT & reference_point, std::vector<LeafConstPtr> & neighbors) const
 {
   neighbors.clear();
@@ -440,7 +462,7 @@ int pclomp::VoxelGridCovariance<PointT>::getNeighborhoodAtPoint1(
 
 //////////////////////////////////////////////////////////////////////////////////////////
 template <typename PointT>
-void pclomp::VoxelGridCovariance<PointT>::getDisplayCloud(
+void autoware::ndt_omp::pclomp::VoxelGridCovariance<PointT>::getDisplayCloud(
   pcl::PointCloud<pcl::PointXYZ> & cell_cloud)
 {
   cell_cloud.clear();
@@ -448,7 +470,7 @@ void pclomp::VoxelGridCovariance<PointT>::getDisplayCloud(
   int pnt_per_cell = 1000;
   boost::mt19937 rng;
   boost::normal_distribution<> nd(0.0, leaf_size_.head(3).norm());
-  boost::variate_generator<boost::mt19937 &, boost::normal_distribution<> > var_nor(rng, nd);
+  boost::variate_generator<boost::mt19937 &, boost::normal_distribution<>> var_nor(rng, nd);
 
   Eigen::LLT<Eigen::Matrix3d> llt_of_cov;
   Eigen::Matrix3d cholesky_decomp;
