+++ package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>ndt_omp</name>
+  <name>autoware_ndt_omp</name>
   <version>0.0.0</version>
   <description>OpenMP boosted NDT and GICP algorithms. With some modifications for Autoware</description>
