+++ src/multigrid_pclomp/multi_voxel_grid_covariance_omp.cpp
@@ -1,6 +1,6 @@
-#include <multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp>
+#include <autoware/ndt_omp/multigrid_pclomp/multi_voxel_grid_covariance_omp_impl.hpp>
 
-#include <multigrid_pclomp/multi_voxel_grid_covariance_omp.h>
+#include <autoware/ndt_omp/multigrid_pclomp/multi_voxel_grid_covariance_omp.h>
 
-template class pclomp::MultiVoxelGridCovariance<pcl::PointXYZ>;
-template class pclomp::MultiVoxelGridCovariance<pcl::PointXYZI>;
+template class autoware::ndt_omp::pclomp::MultiVoxelGridCovariance<pcl::PointXYZ>;
+template class autoware::ndt_omp::pclomp::MultiVoxelGridCovariance<pcl::PointXYZI>;
