+++ src/multigrid_pclomp/multigrid_ndt_omp.cpp
@@ -1,6 +1,8 @@
-#include <multigrid_pclomp/multigrid_ndt_omp_impl.hpp>
+#include <autoware/ndt_omp/multigrid_pclomp/multigrid_ndt_omp_impl.hpp>
 
-#include <multigrid_pclomp/multigrid_ndt_omp.h>
+#include <autoware/ndt_omp/multigrid_pclomp/multigrid_ndt_omp.h>
 
-template class pclomp::MultiGridNormalDistributionsTransform<pcl::PointXYZ, pcl::PointXYZ>;
-template class pclomp::MultiGridNormalDistributionsTransform<pcl::PointXYZI, pcl::PointXYZI>;
+template class autoware::ndt_omp::pclomp::MultiGridNormalDistributionsTransform<
+  pcl::PointXYZ, pcl::PointXYZ>;
+template class autoware::ndt_omp::pclomp::MultiGridNormalDistributionsTransform<
+  pcl::PointXYZI, pcl::PointXYZI>;
