+++ src/pclomp/gicp_omp.cpp
@@ -1,8 +1,10 @@
 // this cannot be swapped
 // clang-format off
-#include <pclomp/gicp_omp.h>
-#include <pclomp/gicp_omp_impl.hpp>
+#include <autoware/ndt_omp/pclomp/gicp_omp.h>
+#include <autoware/ndt_omp/pclomp/gicp_omp_impl.hpp>
 // clang-format on
 
-template class pclomp::GeneralizedIterativeClosestPoint<pcl::PointXYZ, pcl::PointXYZ>;
-template class pclomp::GeneralizedIterativeClosestPoint<pcl::PointXYZI, pcl::PointXYZI>;
+template class autoware::ndt_omp::pclomp::GeneralizedIterativeClosestPoint<
+  pcl::PointXYZ, pcl::PointXYZ>;
+template class autoware::ndt_omp::pclomp::GeneralizedIterativeClosestPoint<
+  pcl::PointXYZI, pcl::PointXYZI>;
